+++ node.go
@@ -237,6 +237,47 @@ func (node *Node) hashWithCount() ([]byte, int64) {
 	return node.hash, hashCount + 1
 }
 
+// validate validates the node contents
+func (node *Node) validate() error {
+	if node == nil {
+		return errors.New("node cannot be nil")
+	}
+	if node.key == nil {
+		return errors.New("key cannot be nil")
+	}
+	if node.version <= 0 {
+		return errors.New("version must be greater than 0")
+	}
+	if node.height < 0 {
+		return errors.New("height cannot be less than 0")
+	}
+	if node.size < 1 {
+		return errors.New("size must be at least 1")
+	}
+
+	if node.height == 0 {
+		// Leaf nodes
+		if node.value == nil {
+			return errors.New("value cannot be nil for leaf node")
+		}
+		if node.leftHash != nil || node.leftNode != nil || node.rightHash != nil || node.rightNode != nil {
+			return errors.New("leaf node cannot have children")
+		}
+		if node.size != 1 {
+			return errors.New("leaf nodes must have size 1")
+		}
+	} else {
+		// Inner nodes
+		if node.value != nil {
+			return errors.New("value must be nil for non-leaf node")
+		}
+		if node.leftHash == nil && node.rightHash == nil {
+			return errors.New("inner node must have children")
+		}
+	}
+	return nil
+}
+
 // Writes the node's hash to the given io.Writer. This function expects
 // child hashes to be already set.
 func (node *Node) writeHashBytes(w io.Writer) error {
@@ -388,17 +429,27 @@ func (node *Node) calcBalance(t *ImmutableTree) int {
 
 // traverse is a wrapper over traverseInRange when we want the whole tree
 func (node *Node) traverse(t *ImmutableTree, ascending bool, cb func(*Node) bool) bool {
-	return node.traverseInRange(t, nil, nil, ascending, false, 0, func(node *Node, depth uint8) bool {
+	return node.traverseInRange(t, nil, nil, ascending, false, 0, false, func(node *Node, depth uint8) bool {
+		return cb(node)
+	})
+}
+
+// traversePost is a wrapper over traverseInRange when we want the whole tree post-order
+func (node *Node) traversePost(t *ImmutableTree, ascending bool, cb func(*Node) bool) bool {
+	return node.traverseInRange(t, nil, nil, ascending, false, 0, true, func(node *Node, depth uint8) bool {
 		return cb(node)
 	})
 }
 
 // nolint:unused,deadcode
 func (node *Node) traverseWithDepth(t *ImmutableTree, ascending bool, cb func(*Node, uint8) bool) bool {
-	return node.traverseInRange(t, nil, nil, ascending, false, 0, cb)
+	return node.traverseInRange(t, nil, nil, ascending, false, 0, false, cb)
 }
 
-func (node *Node) traverseInRange(t *ImmutableTree, start, end []byte, ascending bool, inclusive bool, depth uint8, cb func(*Node, uint8) bool) bool {
+func (node *Node) traverseInRange(t *ImmutableTree, start, end []byte, ascending bool, inclusive bool, depth uint8, post bool, cb func(*Node, uint8) bool) bool {
+	if node == nil {
+		return false
+	}
 	afterStart := start == nil || bytes.Compare(start, node.key) < 0
 	startOrAfter := start == nil || bytes.Compare(start, node.key) <= 0
 	beforeEnd := end == nil || bytes.Compare(node.key, end) < 0
@@ -408,37 +459,46 @@ func (node *Node) traverseInRange(t *ImmutableTree, start, end []byte, ascending
 
 	// Run callback per inner/leaf node.
 	stop := false
-	if !node.isLeaf() || (startOrAfter && beforeEnd) {
+	if !post && (!node.isLeaf() || (startOrAfter && beforeEnd)) {
 		stop = cb(node, depth)
 		if stop {
 			return stop
 		}
 	}
-	if node.isLeaf() {
+
+	if !node.isLeaf() {
+		if ascending {
+			// check lower nodes, then higher
+			if afterStart {
+				stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+			if stop {
+				return stop
+			}
+			if beforeEnd {
+				stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+		} else {
+			// check the higher nodes first
+			if beforeEnd {
+				stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+			if stop {
+				return stop
+			}
+			if afterStart {
+				stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, post, cb)
+			}
+		}
+	}
+	if stop {
 		return stop
 	}
 
-	if ascending {
-		// check lower nodes, then higher
-		if afterStart {
-			stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, cb)
-		}
+	if post && (!node.isLeaf() || (startOrAfter && beforeEnd)) {
+		stop = cb(node, depth)
 		if stop {
 			return stop
-		}
-		if beforeEnd {
-			stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, cb)
-		}
-	} else {
-		// check the higher nodes first
-		if beforeEnd {
-			stop = node.getRightNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, cb)
-		}
-		if stop {
-			return stop
-		}
-		if afterStart {
-			stop = node.getLeftNode(t).traverseInRange(t, start, end, ascending, inclusive, depth+1, cb)
 		}
 	}