+++ node_test.go
@@ -5,6 +5,7 @@ import (
 	"math/rand"
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -29,6 +30,54 @@ func TestNode_aminoSize(t *testing.T) {
 	// non-leaf node
 	node.height = 1
 	require.Equal(t, 57, node.aminoSize())
+}
+
+func TestNode_validate(t *testing.T) {
+	k := []byte("key")
+	v := []byte("value")
+	h := []byte{1, 2, 3}
+	c := &Node{key: []byte("child"), value: []byte("x"), version: 1, size: 1}
+
+	testcases := map[string]struct {
+		node  *Node
+		valid bool
+	}{
+		"nil node":               {nil, false},
+		"leaf":                   {&Node{key: k, value: v, version: 1, size: 1}, true},
+		"leaf with nil key":      {&Node{key: nil, value: v, version: 1, size: 1}, false},
+		"leaf with empty key":    {&Node{key: []byte{}, value: v, version: 1, size: 1}, true},
+		"leaf with nil value":    {&Node{key: k, value: nil, version: 1, size: 1}, false},
+		"leaf with empty value":  {&Node{key: k, value: []byte{}, version: 1, size: 1}, true},
+		"leaf with version 0":    {&Node{key: k, value: v, version: 0, size: 1}, false},
+		"leaf with version -1":   {&Node{key: k, value: v, version: -1, size: 1}, false},
+		"leaf with size 0":       {&Node{key: k, value: v, version: 1, size: 0}, false},
+		"leaf with size 2":       {&Node{key: k, value: v, version: 1, size: 2}, false},
+		"leaf with size -1":      {&Node{key: k, value: v, version: 1, size: -1}, false},
+		"leaf with left hash":    {&Node{key: k, value: v, version: 1, size: 1, leftHash: h}, false},
+		"leaf with left child":   {&Node{key: k, value: v, version: 1, size: 1, leftNode: c}, false},
+		"leaf with right hash":   {&Node{key: k, value: v, version: 1, size: 1, rightNode: c}, false},
+		"leaf with right child":  {&Node{key: k, value: v, version: 1, size: 1, rightNode: c}, false},
+		"inner":                  {&Node{key: k, version: 1, size: 1, height: 1, leftHash: h, rightHash: h}, true},
+		"inner with nil key":     {&Node{key: nil, value: v, version: 1, size: 1, height: 1, leftHash: h, rightHash: h}, false},
+		"inner with value":       {&Node{key: k, value: v, version: 1, size: 1, height: 1, leftHash: h, rightHash: h}, false},
+		"inner with empty value": {&Node{key: k, value: []byte{}, version: 1, size: 1, height: 1, leftHash: h, rightHash: h}, false},
+		"inner with left child":  {&Node{key: k, version: 1, size: 1, height: 1, leftHash: h}, true},
+		"inner with right child": {&Node{key: k, version: 1, size: 1, height: 1, rightHash: h}, true},
+		"inner with no child":    {&Node{key: k, version: 1, size: 1, height: 1}, false},
+		"inner with height 0":    {&Node{key: k, version: 1, size: 1, height: 0, leftHash: h, rightHash: h}, false},
+	}
+
+	for desc, tc := range testcases {
+		tc := tc // appease scopelint
+		t.Run(desc, func(t *testing.T) {
+			err := tc.node.validate()
+			if tc.valid {
+				assert.NoError(t, err)
+			} else {
+				assert.Error(t, err)
+			}
+		})
+	}
 }
 
 func BenchmarkNode_aminoSize(b *testing.B) {